requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
pillow>=8.0.0
numpy>=1.20.0
